fix: use Long for offset to support files larger than 2.14GB
Change offset from Int to Long in IONFILEReadInChunksOptions and
IONFILEReadOptions to prevent integer overflow when reading files
beyond the ~2.14GB Int.MAX_VALUE boundary.

+++ src/test/java/io/ionic/libs/ionfilesystemlib/helper/IONFILELocalFilesHelperTest.kt
@@ -693,7 +693,7 @@ class IONFILELocalFilesHelperTest : IONFILEBaseJUnitTest() {
             fullPath = path,
             options = IONFILEReadOptions(
                 encoding = IONFILEEncoding.WithCharset(Charsets.UTF_8),
-                offset = content.length
+                offset = content.length.toLong()
             )
         )
 
@@ -719,7 +719,7 @@ class IONFILELocalFilesHelperTest : IONFILEBaseJUnitTest() {
             fullPath = path,
             options = IONFILEReadInChunksOptions(
                 encoding = IONFILEEncoding.WithCharset(Charsets.UTF_8),
-                offset = content.length + 100,
+                offset = (content.length + 100).toLong(),
                 chunkSize = 10
             )
         ).test {
